HCI_ERR_KEY_MISSING error possible solution
Under normal conditions, no file can be sent to a non pair or pair device. In case of pairing, when the connection is made and disconnected by the other device in the device selection screen that appears before sending the file, it can be detected by the other device.

After these updates, file transfer can be made to a pair or non-pair device.

Change-Id: I8541d7623d286aaab46cb64ad910a71768c414c0
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>